+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ News
 2.0.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+* Treat strings in the WSGI environment as native strings, compliant with
+  PEP-3333.
 
 
 2.0.2 (2013-03-15)